requirements.txt
fastapi
uvicorn
python-nmap
jinja2
python-dotenv
python-multipart
sqlalchemy